- Added ball stuck to the bar when starting (closes #14 and fixes #21) - Added first try of scoring system, which just counts how many blocks you've broken ATM. - Fixed some incorrect comments.
This makes the game playable and almost done (all to-do list items on issues #3 and #8 are finally checked!). I know there is lots of room to improvements to the code, but this will do at the moment.

Compilation tests: PASS @ Symphony EDA Sonata 3.0
Will test when I have a live board available.

Signed-off-by: jaBote <javote91@hotmail.com>